requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
telethon
sqlalchemy==1.3.23
psycopg2-binary
pytube
youtube_search
youtube_dl
hachoir
ffmpeg 
ffprobe
jikanpy
redis 
rembg 
Pillow
torch==1.8.1+cpu
torchvision==0.9.1+cpu